Save to file deletion log if path is error_new

+++ css_archiving_format.py
@@ -224,22 +224,28 @@ def delete_appraisal_letters(input_dir, df_appraisal):
         if row.in_document_name != '' and row.in_document_name != 'nan':
             name = row.in_document_name
             file_path = update_path(name, input_dir)
-            try:
-                file_deletion_log(log_path, file_path, row.Appraisal_Category)
-                os.remove(file_path)
-            except FileNotFoundError:
-                file_deletion_log(log_path, file_path, 'Cannot delete: FileNotFoundError')
+            if file_path == 'error_new':
+                file_deletion_log(log_path, name, 'Cannot determine file path: new path pattern in metadata')
+            else:
+                try:
+                    file_deletion_log(log_path, file_path, row.Appraisal_Category)
+                    os.remove(file_path)
+                except FileNotFoundError:
+                    file_deletion_log(log_path, file_path, 'Cannot delete: FileNotFoundError')
 
         # Deletes individual letters, not form letters, sent to constituents, if the "out" column isn't blank.
         if row.out_document_name != '' and row.out_document_name != 'nan' and 'form' not in row.out_document_name:
             name = row.out_document_name
             file_path = update_path(name, input_dir)
+            if file_path == 'error_new':
+                file_deletion_log(log_path, name, 'Cannot determine file path: new path pattern in metadata')
             # Only delete if it is a file. Sometimes, out_document_name has the path to a folder instead.
-            if os.path.isfile(file_path):
-                file_deletion_log(log_path, file_path, row.Appraisal_Category)
-                os.remove(file_path)
-            elif not os.path.exists(file_path):
-                file_deletion_log(log_path, file_path, 'Cannot delete: FileNotFoundError')
+            else:
+                if os.path.isfile(file_path):
+                    file_deletion_log(log_path, file_path, row.Appraisal_Category)
+                    os.remove(file_path)
+                elif not os.path.exists(file_path):
+                    file_deletion_log(log_path, file_path, 'Cannot delete: FileNotFoundError')
 
 
 def file_deletion_log(log_path, file_path, note):
